# ohbs-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/webhooks": {
      "post": {
        "operationId": "WebhooksController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterWebhookEndpointDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "[Admin] Registrar endpoint webhook — secreto de firma se muestra una vez",
        "tags": [
          "webhooks"
        ]
      },
      "get": {
        "operationId": "WebhooksController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "[Admin] Listar endpoints webhook",
        "tags": [
          "webhooks"
        ]
      }
    },
    "/api/v1/webhooks/{id}/deactivate": {
      "post": {
        "operationId": "WebhooksController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "[Admin] Desactivar endpoint webhook",
        "tags": [
          "webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "CmorFlow Tax API",
    "description": "Plataforma tributaria B2B para facturación electrónica chilena vía API. Emisión asíncrona de DTE, RCOF, webhooks y artefactos XML/PDF. Autenticación HMAC por credencial ligada a un único tenant.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "integration-hmac": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Autenticación HMAC. Requiere X-Timestamp, X-Nonce y X-Signature. Credenciales API: cmor_live_*, admin: cmor_admin_*."
      }
    },
    "schemas": {
...
      "RegisterWebhookEndpointDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://api.cmorapr.cl/hooks/cmorflow"
          },
          "events": {
            "type": "string",
            "example": [
              "dte.accepted",
              "dte.rejected"
            ],
            "enum": [
              "dte.submitted",
              "dte.accepted",
              "dte.observed",
              "dte.rejected",
              "dte.failed",
              "rcof.submitted",
              "rcof.accepted",
              "rcof.observed",
              "rcof.rejected",
              "rcof.failed"
            ]
          },
          "description": {
            "type": "string",
            "example": "Producción CMORAPR"
          }
        },
        "required": [
          "url",
          "events"
        ]
      }
    }
  }
}